+++ libs/safe/src/lib/components/ui/aggregation-builder/pipeline/add-field-stage/add-field-stage.stories.ts
@@ -80,6 +80,9 @@ const TEMPLATE: StoryFn<SafeAddFieldStageComponent> = () => ({
   },
 });
 
+/**
+ * Default story.
+ */
 export const DEFAULT = {
   render: TEMPLATE,
   name: 'Default',