+++ libs/safe/src/lib/components/widgets/map-settings/map-properties/webmap-select/webmap-select.component.ts
@@ -54,6 +54,7 @@ export class WebmapSelectComponent implements ControlValueAccessor, OnInit {
    * Map Properties of Map widget.
    *
    * @param arcgis service
+   * @param ngControl current ng control
    */
   constructor(
     private arcgis: ArcgisService,
@@ -87,7 +88,6 @@ export class WebmapSelectComponent implements ControlValueAccessor, OnInit {
   public selectionOnChange(e: any) {
     this.value = e.value;
     this.onChanged(this.value);
-    console.log(this.ngControl.value);
   }
 
   /**
@@ -123,6 +123,8 @@ export class WebmapSelectComponent implements ControlValueAccessor, OnInit {
 
   /**
    * Search for webmap data in argcis-rest-request using arcgis service
+   *
+   * @param text search text
    */
   private search(text?: string): void {
     this.arcgis