+++ libs/safe/src/lib/services/map/arcgis.service.ts
@@ -14,6 +14,7 @@ import {
 import * as EsriHeat from 'esri-leaflet-heatmap';
 import * as EsriCluster from 'esri-leaflet-cluster';
 import * as EsriRenderers from 'esri-leaflet-renderers';
+// import * as EsriRenderers from '@oort-front/esri';
 import * as Esri from 'esri-leaflet';
 import * as Vector from 'esri-leaflet-vector';
 import * as Geocoding from 'esri-leaflet-geocoder';
@@ -61,10 +62,7 @@ export class ArcgisService {
    */
   public loadWebMap(
     map: L.Map,
-    id: string,
-    options: {
-      loadBasemap?: boolean;
-    } = { loadBasemap: true }
+    id: string
   ): Promise<{ basemaps: TreeObject[]; layers: TreeObject[] }> {
     return new Promise((resolve) => {
       getItemData(id, {
@@ -72,9 +70,7 @@ export class ArcgisService {
       }).then((webMap: any) => {
         this.setDefaultView(map, webMap);
         Promise.all([
-          options.loadBasemap
-            ? this.loadBaseMap(map, webMap)
-            : Promise.resolve([]),
+          this.loadBaseMap(map, webMap),
           this.loadOperationalLayers(map, webMap),
         ]).then(([basemaps, layers]) => {
           resolve({ basemaps, layers });
@@ -207,6 +203,7 @@ export class ArcgisService {
   ): Promise<TreeObject[]> {
     const layers: TreeObject[] = [];
     for (const layer of webMap.operationalLayers) {
+      console.log(layer);
       await this.addLayer(map, layer, layers);
     }
     return layers;
@@ -267,6 +264,13 @@ export class ArcgisService {
       case 'ArcGISFeatureLayer': {
         let featureLayer: L.Layer | undefined = undefined;
         if (layer.url) {
+          if (layer.itemId) {
+            await getItemData(layer.itemId, {
+              authentication: this.session,
+            }).then((item: any) => {
+              console.log(item);
+            });
+          }
           const reduction = get(layer, 'layerDefinition.featureReduction');
           if (reduction) {
             switch (reduction.type) {
@@ -277,15 +281,18 @@ export class ArcgisService {
                   url: layer.url,
                   token: this.esriApiKey,
                   // If popup is not disabled we want to show the cluster popup, no zoom to bounds
-                  zoomToBoundsOnClick:
-                    layer.layerDefinition.featureReduction.disablePopup,
+                  zoomToBoundsOnClick: get(
+                    layer,
+                    'layerDefinition.featureReduction.disablePopup',
+                    true
+                  ),
                   // opacity,
                   // polygonOptions: {
                   //   opacity,
                   // },
                   // pane,
                   clusterPane: pane,
-                  drawingInfo: layer.layerDefinition.drawingInfo,
+                  drawingInfo: get(layer, 'layerDefinition.drawingInfo'),
                   minZoom: minScaleLevel,
                   maxZoom: maxScaleLevel,
                 });
@@ -349,10 +356,12 @@ export class ArcgisService {
                       });
                     });
                     featureLayer.onAdd = (map: L.Map) => {
+                      // call default onAdd
                       const l = EsriHeat.FeatureLayer.prototype.onAdd.call(
                         featureLayer,
                         map
                       );
+                      // Create legend element
                       const container = document.createElement('div');
                       container.className = 'flex gap-1';
                       const linearGradient = document.createElement('div');
@@ -369,9 +378,23 @@ export class ArcgisService {
                       const html =
                         `<div class="font-bold">${layer.title}</div>` +
                         container.outerHTML;
+                      // Add legend
                       const legendControl = (map as any).legendControl;
                       if (legendControl) {
                         legendControl.addLayer(featureLayer, html);
+                      }
+                      return l;
+                    };
+                    featureLayer.onRemove = (map: L.Map) => {
+                      // call default onRemove
+                      const l = EsriHeat.FeatureLayer.prototype.onRemove.call(
+                        featureLayer,
+                        map
+                      );
+                      // remove legend
+                      const legendControl = (map as any).legendControl;
+                      if (legendControl) {
+                        legendControl.removeLayer(featureLayer);
                       }
                       return l;
                     };
@@ -428,6 +451,7 @@ export class ArcgisService {
             }
           }
         }
+        console.log(featureLayer);
         if (featureLayer) {
           if (get(layer, 'visibility', true) && visibility) {
             featureLayer.addTo(map);
@@ -490,10 +514,22 @@ export class ArcgisService {
           token: this.esriApiKey,
           opacity,
         });
-        layers.push({
-          label: layer.title,
-          layer: tiledMapLayer,
-        });
+        if (layer.itemId) {
+          await getItemData(layer.itemId, {
+            authentication: this.session,
+          }).then((item: any) => {
+            console.log(item);
+          });
+        }
+        if (tiledMapLayer) {
+          if (get(layer, 'visibility', true) && visibility) {
+            tiledMapLayer.addTo(map);
+          }
+          layers.push({
+            label: layer.title,
+            layer: tiledMapLayer,
+          });
+        }
         break;
       }
       case 'WMS': {
@@ -558,6 +594,12 @@ export class ArcgisService {
     return this.http.get(path + `?token=${this.esriApiKey}`).toPromise();
   }
 
+  /**
+   * Execute reverse geosearch
+   *
+   * @param latlng latlng to reverse search
+   * @returns seerch as promise
+   */
   public reverseSearch(latlng: L.LatLng) {
     return new Promise((resolve, reject) =>
       (Geocoding as any)